snowflake_automation/requirements.txt
snowflake-connector-python>=3.0.0
pandas
snowflake-snowpark-python[pandas]
streamlit
